bin/requirements.txt
gitpython
intelhex
parse
pycrypto
pynrfjprog
pyserial
